git-svn-id: file:///space/source.marketcetera.org/root/branches/will@10874 501a4a96-914b-0410-8454-f5988a4ca596

+++ options/source/photon/plugins/org.marketcetera.photon.marketdata/src/main/java/org/marketcetera/photon/internal/marketdata/LatestTickManager.java
@@ -103,7 +103,7 @@ public class LatestTickManager extends
                         }
                         if (data instanceof OptionEvent) {
                             OptionEvent optionData = (OptionEvent) data;
-                            Integer multiplier = optionData.getMultiplier();
+                            BigDecimal multiplier = optionData.getMultiplier();
                             if (multiplier != null) {
                                 item.setMultiplier(multiplier);
                             }
